Move exceptions into their own library
